STYLE: avoid DeprecationWarning messages from vtk wrapping
When passing a float to a vtk method that expects int, a
DeprecationWarning is issued, but it does not provide any
diagnositic information about the method or class that caused
the issue. To get the information, you need to issue the following
python commands before the warning is triggered:

 import warnings
 warnings.simplefilter('error')

This generates a stack trace that can be used to find the issue.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=21294
git-svn-id: http://svn.slicer.org/Slicer4/trunk@21294 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/SlicerApp/Testing/Python/RSNA2012Vis.py
@@ -243,8 +243,8 @@ class RSNA2012VisTest(unittest.TestCase):
     down()
     for step in xrange(steps):
       frac = float(step)/steps
-      x = start[0] + frac*(end[0]-start[0])
-      y = start[1] + frac*(end[1]-start[1])
+      x = int(start[0] + frac*(end[0]-start[0]))
+      y = int(start[1] + frac*(end[1]-start[1]))
       interator.SetEventPosition(x,y)
       style.OnMouseMove()
     up()
